Staging

+++ project/business-logic-layer/i-on-web-auth.js
@@ -7,6 +7,9 @@ const internalErrors = require('../common/i-on-web-errors.js');
 
 const FileStore = require('session-file-store')(session); 
 
+/// Session expiration time
+const sessionMaxAge = 7 * 24 * 60 * 60 * 1000; /// Corresponds to 7 days in milliseconds
+
 module.exports = (app, data, sessionDB) => {
 
 	function userToRef(user, done) {
@@ -28,7 +31,8 @@ module.exports = (app, data, sessionDB) => {
     app.use(session({
 		resave: false,              
 		saveUninitialized: false,  
-		secret: 'secret',   // TO DO - Generate random string
+		secret: 'secret',
+		cookie: { maxAge: sessionMaxAge },
 		store: new FileStore() 
     }))
 
@@ -40,33 +44,35 @@ module.exports = (app, data, sessionDB) => {
     	
     return {
 
-		getAuthMethodsAndFeatures: async function() {
-			return data.loadAuthenticationMethodsAndFeatures();
-        },
-
 		submitInstitutionalEmail: async function(email) {
+
+			/// Using regular expressions to validate email
+			const re = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+			if(!email || !re.test(email)) throw internalErrors.BAD_REQUEST;
+			
 			return data.submitInstitutionalEmail(email);
         }, 
 
 		pollingCore: async function(req, authForPoll) {
-			const receivedTokens = await data.pollingCore(authForPoll);
+			const pollingResponse = await data.pollingCore(authForPoll);
 
 			/// Check if pooling succeeded
-			if(receivedTokens.hasOwnProperty("access_token")) {
+			if(pollingResponse.hasOwnProperty("access_token")) {
 
-				const tokens = receivedTokens.id_token.split(".");
+				/// Obtaining user email
+				const tokens = pollingResponse.id_token.split(".");
 				const user_email = jwt_decode(tokens[1], { header: true }).email;
 				
-				const user = await data.loadUser(receivedTokens.access_token, receivedTokens.token_type, user_email);
-				const sessionId = await sessionDB.createUserSession(user_email, receivedTokens);
+				const user = await data.loadUser(pollingResponse.access_token, pollingResponse.token_type, user_email);
+				const sessionId = await sessionDB.createUserSession(user_email, pollingResponse);
 				
 				const userSessionInfo = Object.assign(
 					{'sessionId': sessionId},
 					user,
-					receivedTokens
+					pollingResponse
 				);
 				
-				/// If the user doesn't have a username, we give one by default. 
+				/// If the user doesn't have a username, we provide one by default. 
 				if(!userSessionInfo.username) {
 					const newUsername = user.email.slice(0, user.email.indexOf("@"));
 					userSessionInfo.username = newUsername;
@@ -85,25 +91,29 @@ module.exports = (app, data, sessionDB) => {
 			}
 		},
 		
-		logout: async function(req) { /// TODO: Verify if user is authenticated and handling errors
-			const user = req.user;
-			
-			req.logout();
-			req.session.destroy(err => { /// TODO : replace ...
-				if (err) {
-					throw internalErrors.SERVICE_FAILURE;
-				}
-			})
+		logout: async function(req) {
+			if(req.user) {
+				const user = req.user;
+				
+				req.logout();
+				req.session.destroy(err => {
+					if (err) {
+						throw internalErrors.SERVICE_FAILURE;
+					}
+				})
 
-			await data.revokeAccessToken(user);
-			await sessionDB.deleteUserSession(user.sessionId);
+				await data.revokeAccessToken(user);
+				await sessionDB.deleteUserSession(user.sessionId);
+			} else {
+				throw internalErrors.UNAUTHENTICATED;
+			}
 		},
 
 		deleteUser: async function(req) {
 			const user = JSON.parse(JSON.stringify(req.user));;
 
 			req.logout();
-			req.session.destroy(err => { /// TODO : replace ...
+			req.session.destroy(err => {
 				if (err) {
 					throw internalErrors.SERVICE_FAILURE;
 				}
@@ -119,8 +129,8 @@ module.exports = (app, data, sessionDB) => {
 /******* Helper functions *******/
 
 const getUserAndSessionInfo = async function(data, sessionDB, sessionId) { // Through the session identifier we will obtain information about the user as well as the session 
-		/// Obtaining user session info from elasticsearch db
-		const sessionInfo = await sessionDB.getUserTokens(sessionId);
+	/// Obtaining user session info from elasticsearch db
+	const sessionInfo = await sessionDB.getUserTokens(sessionId);
 	
 	try { 
 		const userProfileInfo = await data.loadUser(sessionInfo.access_token, sessionInfo.token_type, sessionInfo.email);
@@ -147,7 +157,7 @@ const getUserAndSessionInfo = async function(data, sessionDB, sessionId) { // Th
 	}
 }
 
-const updateUserSession = async function(data, sessionDB, sessionInfo, sessionId) {// mudar assinatura pra receber o session id e talvez email
+const updateUserSession = async function(data, sessionDB, sessionInfo, sessionId) {
 	const newTokens = await data.refreshAccessToken(sessionInfo);
 	await sessionDB.storeUpdatedInfo(sessionInfo.email, newTokens, sessionId)
 }
